Properly modularize the project with maven
