+++ FredBoat/src/main/java/fredboat/Config.java
@@ -26,8 +26,8 @@
 package fredboat;
 
 import com.mashape.unirest.http.exceptions.UnirestException;
+import fredboat.shared.constant.DistributionEnum;
 import fredboat.util.DiscordUtil;
-import fredboat.util.constant.DistributionEnum;
 import org.apache.commons.io.FileUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
