+++ FredBoat/src/main/java/fredboat/audio/MusicPersistenceHandler.java
@@ -34,8 +34,8 @@ import fredboat.audio.queue.AudioTrackContext;
 import fredboat.audio.queue.RepeatMode;
 import fredboat.audio.queue.SplitAudioTrackContext;
 import fredboat.feature.I18n;
-import fredboat.util.constant.DistributionEnum;
-import fredboat.util.constant.ExitCodes;
+import fredboat.shared.constant.DistributionEnum;
+import fredboat.shared.constant.ExitCodes;
 import net.dv8tion.jda.core.entities.Member;
 import net.dv8tion.jda.core.entities.TextChannel;
 import net.dv8tion.jda.core.entities.VoiceChannel;
