+++ FredBoat/src/main/java/fredboat/command/admin/BotRestartCommand.java
@@ -30,7 +30,7 @@ import fredboat.commandmeta.abs.Command;
 import fredboat.commandmeta.abs.ICommand;
 import fredboat.commandmeta.abs.ICommandRestricted;
 import fredboat.perms.PermissionLevel;
-import fredboat.util.constant.ExitCodes;
+import fredboat.shared.constant.ExitCodes;
 import fredboat.util.TextUtils;
 import net.dv8tion.jda.core.entities.Guild;
 import net.dv8tion.jda.core.entities.Member;
