+++ FredBoat/src/main/java/fredboat/command/moderation/PermissionsCommand.java
@@ -35,9 +35,9 @@ import fredboat.feature.I18n;
 import fredboat.feature.togglz.FeatureFlags;
 import fredboat.perms.PermissionLevel;
 import fredboat.perms.PermsUtil;
+import fredboat.shared.constant.BotConstants;
 import fredboat.util.ArgumentUtil;
 import fredboat.util.TextUtils;
-import fredboat.util.constant.BotConstants;
 import net.dv8tion.jda.core.EmbedBuilder;
 import net.dv8tion.jda.core.Permission;
 import net.dv8tion.jda.core.entities.Guild;
