+++ FredBoat/src/main/java/fredboat/commandmeta/CommandManager.java
@@ -34,10 +34,10 @@ import fredboat.commandmeta.abs.IMusicCommand;
 import fredboat.feature.I18n;
 import fredboat.perms.PermissionLevel;
 import fredboat.perms.PermsUtil;
+import fredboat.shared.constant.BotConstants;
+import fredboat.shared.constant.DistributionEnum;
 import fredboat.util.DiscordUtil;
 import fredboat.util.TextUtils;
-import fredboat.util.constant.BotConstants;
-import fredboat.util.constant.DistributionEnum;
 import fredboat.util.rest.RestActionScheduler;
 import net.dv8tion.jda.core.Permission;
 import net.dv8tion.jda.core.entities.Guild;
