+++ Shared/src/main/java/fredboat/shared/constant/BotConstants.java
@@ -23,7 +23,7 @@
  *
  */
 
-package fredboat.util.constant;
+package fredboat.shared.constant;
 
 import java.awt.*;
